Flattened `commit`.
